refactor: use enhanced forms
